Frame Work back end
Made many changes to begin the real work. Note that I changed the
database Version to dateTime in stead of byte ... that was wrong in the
data base setup.

1. Changed the repositories but will have a few more changes
2. Fixed an issue i was having with login
3. Added two viewmodels for the code base
4. Added the changes to the admin view

5. Need to implement the view model for the admin view and this part is
done (10 mins worth of work)

Need to finish the Version creation method in a new controler and call
it from describe screens in sdd but most of frame work is done now.